Move plugin config into its own file.
